Modified Command/GameState interaction
Added a few GameState methods that allow Commands to directly access the
Player and Room objects. Then Commands can call those mutator methods
directly instead of calling GameState interface methods. I figured we
didn't need more layers of indirection here. We just want the Commands
to change things directly. If we want to change how the game engine acts
in the future, it can make sense that we have to change with the
commands do.