Added CommandTest testAddHealth.
This relies on the specification that player health is not enforced
(capped) by max health. We may want to rethink this in the future. If we
make it so health caps at max health, we need to change this unit test,
and we should expect it to fail under certain circumstances.