Added GameModel.look()
This brings functionality out of Command so that it can be changed later
in a place that makes more sense (I think). Also this standardizes
'look' for GameMain and Command. GameMain still has the option to grab
these descriptions independently (like when entering a room for the
first time), but the default behavior of 'look' is defined here in
GameMain.