Implemented Command.parseScript()
Now Command should be finished. You should be able to run Commands and
they will make the appropriate calls.

The trickiest part here is Commands that contain whole Scripts in their
arguments. ADD_SCRIPT and ADD_SCRIPT_TO Commands need to parse the
arguments and build a Script object to add to some room in the
gameState.

I also moved removeScript and removeScriptFrom back into the switch
statement. It didn't make sense to break them out into separate methods
when we don't do it for any of the other ones.